Rename package from gateway to controller
